two separate packages for lower and upper network levels
