+++ satviz-consumer/src/main/java/edu/kit/satviz/consumer/ConsumerApplication.java
@@ -20,9 +20,9 @@ import edu.kit.satviz.consumer.processing.Mediator;
 import edu.kit.satviz.consumer.processing.RecencyHeatmap;
 import edu.kit.satviz.consumer.processing.RingInteractionGraph;
 import edu.kit.satviz.consumer.processing.VariableInteractionGraph;
-import edu.kit.satviz.network.ConsumerConnection;
-import edu.kit.satviz.network.OfferType;
-import edu.kit.satviz.network.ProducerId;
+import edu.kit.satviz.network.pub.ConsumerConnection;
+import edu.kit.satviz.network.pub.OfferType;
+import edu.kit.satviz.network.pub.ProducerId;
 import edu.kit.satviz.parsers.DimacsFile;
 import edu.kit.satviz.parsers.ParsingException;
 import edu.kit.satviz.sat.ClauseUpdate;
