+++ satviz-consumer/src/main/java/edu/kit/satviz/consumer/processing/Mediator.java
@@ -4,9 +4,9 @@ import edu.kit.satviz.consumer.config.ConsumerConfig;
 import edu.kit.satviz.consumer.config.WeightFactor;
 import edu.kit.satviz.consumer.display.VideoController;
 import edu.kit.satviz.consumer.graph.Graph;
-import edu.kit.satviz.network.ConsumerConnection;
-import edu.kit.satviz.network.ConsumerConnectionListener;
-import edu.kit.satviz.network.ProducerId;
+import edu.kit.satviz.network.pub.ConsumerConnection;
+import edu.kit.satviz.network.pub.ConsumerConnectionListener;
+import edu.kit.satviz.network.pub.ProducerId;
 import edu.kit.satviz.sat.ClauseUpdate;
 import edu.kit.satviz.sat.SatAssignment;
 import edu.kit.satviz.serial.SerializationException;
@@ -14,7 +14,6 @@ import java.io.IOException;
 import java.util.Queue;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 import javafx.scene.paint.Color;
