+++ satviz-network/src/main/java/edu/kit/satviz/network/general/ClientConnectionManager.java
@@ -1,4 +1,4 @@
-package edu.kit.satviz.network;
+package edu.kit.satviz.network.general;
 
 import java.io.IOException;
 import java.net.InetSocketAddress;
