+++ satviz-network/src/main/java/edu/kit/satviz/network/general/ConnectionContext.java
@@ -1,4 +1,4 @@
-package edu.kit.satviz.network;
+package edu.kit.satviz.network.general;
 
 import java.io.IOException;
 import java.net.ConnectException;
