+++ satviz-network/src/main/java/edu/kit/satviz/network/pub/ConsumerConnection.java
@@ -1,5 +1,8 @@
-package edu.kit.satviz.network;
+package edu.kit.satviz.network.pub;
 
+import edu.kit.satviz.network.general.ConnectionId;
+import edu.kit.satviz.network.general.NetworkMessage;
+import edu.kit.satviz.network.general.ServerConnectionManager;
 import edu.kit.satviz.sat.Clause;
 import edu.kit.satviz.sat.ClauseUpdate;
 import edu.kit.satviz.sat.SatAssignment;
