+++ satviz-network/src/main/java/module-info.java
@@ -4,4 +4,6 @@ module edu.kit.satviz.network {
   requires edu.kit.satviz.serial;
 
   exports edu.kit.satviz.network;
+    exports edu.kit.satviz.network.general;
+    exports edu.kit.satviz.network.pub;
 }
