+++ satviz-network/src/test/java/edu/kit/satviz/network/ProducerConsumerTest.java
@@ -1,5 +1,6 @@
 package edu.kit.satviz.network;
 
+import edu.kit.satviz.network.pub.*;
 import edu.kit.satviz.sat.Clause;
 import edu.kit.satviz.sat.ClauseUpdate;
 import edu.kit.satviz.sat.SatAssignment;
