+++ satviz-producer/src/main/java/edu/kit/satviz/producer/mode/SolverMode.java
@@ -4,8 +4,8 @@ import edu.kit.ipasir4j.Ipasir;
 import edu.kit.ipasir4j.IpasirNotFoundException;
 import edu.kit.ipasir4j.Solver;
 import edu.kit.satviz.common.Hashing;
-import edu.kit.satviz.network.OfferType;
-import edu.kit.satviz.network.ProducerId;
+import edu.kit.satviz.network.pub.OfferType;
+import edu.kit.satviz.network.pub.ProducerId;
 import edu.kit.satviz.parsers.DimacsFile;
 import edu.kit.satviz.parsers.ParsingException;
 import edu.kit.satviz.producer.ProducerMode;
